Reject reservation on insufficient capacity
Again, using the Devil's Advocate technique, it'd definitely be possible
to come up with an incorrect implementation that would nonetheless pass
all tests. Since this isn't the purpose of this code base, at this time,
the implementation here is correct according to the intent of the two
test cases present.