Move configuration of capacity out
This will decouple the CompositionRoot class itself from the
Startup file, which could come in handy if one were to engage in
some in-process hosting for integration testing purposes. That's not a
plan for this code base, though, but this is an implementation pattern
that has served me well in the past.