Add SqlReservationsRepository
To keep this example code base simple, there's no tests of this
implementation. One could hand-wave the ommission away by calling it a
Humble Object (see Meszaros 2007), but it'd be more appropriate to cover
this with some integration tests. I've plenty of examples of that in
other places, so will not include that here, as it adds complexity.