Address CQS violation in IReservationsRepository
This doesn't change the CQS violation in IMaîtreD, but it's a step in
that direction.

Notice that this innocuous changed prompted a change in a test, because
a Stub was configured to replay the interaction with the database. This
could be seen as a symptom of Overspecified Software, or, in general,
tests that are too brittle.