Use concrete MaîtreD in ReservationsController
The MaîtreD class is now pure. Its deterministic behaviour enables us to
completely control it from unit tests, so there's no reason to hide it
behind an interface.

I renamed one of the tests because the test case subtly changes. When
previously the test used an IMaîtreD Test Double, it configured the Test
Double to accept the reservation, thereby implying that there was enough
remaining capacity, regardless of the number of existing reservations.

Now that ReservationsController uses the real MaîtreD class, the
existing reservations as returned by the repository determines whether
or not the reservation is accepted. Since there's no explicit setup for
the repository Test Double's ReadReservations method, it'll return a
default instance of IEnumerable<Reservation>, which effectively
corresponds to an empty collection. So, implicitly, from here going
forward, the test case is that there's no prior reservations.