Use concrete Mapper
Like the previous, similar refactorings, the Map method on Mapper is a
pure function, which means that it's controllable from unit tests.

Unfortunately, that Reservation remains a mutable object without
structural equality occasions awkward Mock setup. The creation of an
entire ReservationComparer class may seem overkill to address a single
issue, but my experience tells me that if I can make a class immutable,
I may as well add a comparer as soon as I run into an issue like this,
because it'll turn out to be reusable later.