Parametrise a test
When the test was a fact, I could use the Devil's Advocate technique to
subtly change the TryAccept implementation. Instead of a 'less than'
comparison, a 'less than or equal' comparison also passed all tests:

if (Capacity <= reservedSeats + reservation.Quantity)
    return null;

While this was a change I deliberately made to demonstrate that the
tests didn't provide enough coverage, it's a kind of change one could
easily make inadvertently.

Apart from that, one should notice that it's much harder to apply the
Devil's Advocate technique with the state-based testing strategy shown
here, compared to the interaction-based testing strategy previously
showcased. This is because the ReservationsControllerTests effectively
add extra test cases that also cover MaîtreD.TryAccept and
Validator.Validate.