Parametrise ValidatorTests
This is mostly just in order to complete the refactoring from
single-test-case test methods (facts, in xUnit.net parlance) to
parametrised tests (theories).

Please note that it's entirely acceptable to have a mix of facts and
theories in a test suite. The intention with this commit, then, isn't to
imply that all tests should be parametrised, but only to demonstrate
that in this code base, it's possible to do so.