Promote existing reservations to test argument
One circumstance that could prompt this is another Devil's Advocate
change to MaîtreD. While perhaps not something you might do
inadvertently, I could change the branch in TryAccept to the following
without breaking any tests:

if (reservations.Any() || Capacity < reservation.Quantity)
    return null;

The added test case flushes out the appropriate implementation by
preventing that particular Devil's Advocate change.

Notice that the only thing that matters so far is the Quantity values
for each Reservation. That's a bit odd, so ought to be addressed in the
future, but I'll leave it as such for now.