Turn test into FsCheck property
The capacity supplied by FsCheck is an arbitrary integer, including
negative values and zero. A future commit will narrow the scope.

You can see the test fail by commenting out the branch that returns a
BadRequest result in the ReservationsController. This will cause the
Post method to throw an exception, so that the assertions are never
executed. To get around that, you can supply a proper date string to the
dto variable. This will cause the assertion to fail.