Reframe test argument to capacity surplus
As a preliminary step to turning this test into a property, it can help
reframing the test arguments. Notice that before the change, the
capacity value was, in all test cases, greater than or equal to the
requested quantity.

Instead of defining the capacity as an absolute value, this change turns
it into a relative value. The capacitySurplus represents the margin of
capacity, above the requested quantity. Notice that for one of the test
cases, the margin is 0.