Turn no-prior-reservations test into property
From the previous commit's reframing of the problem, it should be clear
that the capacity surplus can be any non-negative integer, whereas the
quantity can be any positive integer (i.e. 0 isn't a valid value for
that argument).

To be entirely honest, issues with integer overflows are conceivable
when values are constrained to 32-bit integers, 64-bit integers, and so
on. If, for instance, the surplus is the maximum 32-bit value, adding any
positive number to it would cause it to overflow.

In practice, I've never seen FsCheck do this with PositiveInt,
NonNegativeInt, and like types. I admit that I haven't researched why
that tends not to be a problem, but I'm guessing that the generators
associated with those types steer clear of extremes. I'll leave that
research as an exercise to the reader :)

Just to underscore the point, I tried to change the property declaration
to:

[Property(MaxTest = 1_000_000)]

Even with a million repetitions, the property still passes.