Reframe quantity test argument to excessQuantity
This is, overall, the same gambit as a few commits ago (q.v.), but with
a variation. Instead of changing the capacity to a, say,
capacityDeficit, I change the quantity to excess quantity. It turns out
that this holds better potential of being turned into a property.

It's not that I could predict this already. The real story is that I
tried to make the change where I reframed the test into one of a
capacity deficit and a quantity, whereafter I turned it into an FsCheck
property.

When I ran the property, however, it turned out that it generated
negative (or 0) capacity values, which are disallowed by the
ReservationsController constructor.

So I reset the Git repository and made this change instead.