Turn sold-out test case into property
Based on the previous commit, it's straightforward to turn the test
method into a property. It's clear from the constraints imposed by the
ReservationsController constructor that capacity must be a positive
integer, and that the excess quantity requested must be at least one
(i.e. another positive integer), because otherwise it wouldn't be in
excess.