Feed property with full list of prior reservations
Until now, the property used a single Reservation object with a Quantity
fed from the property as a stand-in for an arbitrary number of
reservations that would sum to the number of reserved seats. This change
widens the scope of the test considerable by feeding the property with
an arbitrarily-sized list of Reservation objects (including empty
lists).

This is in my opinion one of the killer benefits of property-based
testing with a library like FsCheck. Granted, all of this can be written
in a nicer way in F#.

It should now be clearer why I initially decided to let the number of
reserved seats be the single absolute value from which other values
could be derived. I knew that I'd like to arrive at this sort of test,
and when you generate a list, the number of elements is an absolute
number.