Turn parametrised test into property
This is the final test converted into a property. Again, you don't have
to go all in on property-based testing. It's perfectly fine to mix
facts, theories, and properties in the same test suite.

The reason I converted all the test methods was mainly to demonstrate
that you can turn most tests into properties if you'd like. It's my
experience that this technique applies widely.

For an example, however, of a problem where, it seems to me at least,
at least some concrete test data are required, consider the Roman
Numerals kata:
https://blog.ploeh.dk/2016/06/28/roman-numerals-via-property-based-tdd