Remove test code
This commit comes with 5 FAILING TESTS. That's on purpose. The exercise
is to make those 5 tests pass by properly defining the expected
interaction between the SUT and its dependencies.

It's probably easiest to start with the MaïtreDTests, and then
subsequently proceed to the ReservationsControllerTests.

You can use Moq for Test Doubles, but that's not required. Another
option is to write Test Spies and Stubs by hand. Moq is already,
however, available in the test project.